requirements.txt
aiogram==3.13.1
APScheduler==3.10.4
python-dotenv==1.0.1
pytz==2024.1
aiosqlite==0.20.0
fastapi==0.115.0
uvicorn==0.30.6
